teste3

+++ ecommerce/src/main/java/br/edu/unifio/ecommerce/repositorios/PedidoRepositorio.java
@@ -3,6 +3,6 @@ package br.edu.unifio.ecommerce.repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.unifio.ecommerce.entidades.Pedido;
 
-public interface PedidoRepositorio extends JpaRepository<Pedido, Short> {
+public interface PedidoRepositorio extends JpaRepository<Pedido, Integer> {
 
 }
